Fix typecheck errors
